requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
defusedxml==0.6.0
Django==3.0
django-braces==1.14.0
django-cors-headers==3.3.0
django-filter==2.2.0
django-oauth-toolkit==1.3.2
django-pagedown==2.1.2
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.11.0
drf-nested-routers==0.91
idna==2.9
importlib-metadata==1.6.0
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
Pillow==7.1.2
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
simplejson==3.17.0
six==1.15.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
zipp==3.1.0
